% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic8ball.R
\name{magic8ball}
\alias{magic8ball}
\title{Magic 8 ball function that return you an answer to your question}
\usage{
magic8ball(question)
}
\arguments{
\item{question}{character ask any yes/no question!}
}
\value{
character the answer to your question
}
\description{
Magic 8 ball function that return you an answer to your question
}
